fix(lexical-editor): refactor image nodes
