+++ packages/lexical-editor/src/core/nodes/cms.ts
@@ -2,6 +2,5 @@ import {basicEditorNodes} from './basic'
 
 export const cmsEditorNodes = [
   ...basicEditorNodes,
-  // ImageNode,
   // EmbedNode,
 ]
