+++ packages/lexical-editor/src/react/plugins/ImagePlugin/Image.scss
@@ -27,7 +27,7 @@
 
   @include mq.desktop-and-above {
     position: relative;
-    margin-left: auto;
+    float: right;
 
     &:after {
       width: 100%;
