+++ packages/lexical-editor/src/react/plugins/ImagePlugin/nodes/index.tsx
@@ -1,9 +1,7 @@
 import { ImageNode } from './ImageNode'
-import { ImageContentNode } from './ImageContentNode'
 
-export { ImageNode, ImageContentNode }
+export { ImageNode }
 
 export default {
   ImageNode,
-  ImageContentNode,
 }
